pyproject.toml
[project]
name = "dailydo-todo-app"
version = "0.1.0"
description = ""
authors = [
    {name = "M-MuneebAlam",email = "mmuneeb.alam09@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "psycopg[binary] (>=3.2.5,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
